refactor: ディレクト構成の変更
